# docs/source/reference/api-restapi/openapi.yml
openapi: 3.0.1
info:
  title: Radar REST API
  version: 0.0.0
servers:
- url: /
...
paths:
...
  /api/v1/artifacts/{id}/snapshots:
    get:
      tags:
      - Artifacts
      summary: Gets the Snapshots for the resource
      operationId: get_resource_snapshots_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the artifact resource.
        required: true
        schema:
          type: integer
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsArtifactPage'
  /api/v1/artifacts/{id}/snapshots/{snapshotId}:
    get:
      tags:
      - Artifacts
      summary: Gets a Snapshot for the resource by snapshot id
      operationId: get_resources_snapshots_id_endpoint
      parameters:
      - name: snapshotId
        in: path
        description: Snapshot ID for the artifact resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the artifact resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsIdArtifact-dump'
...
  /api/v1/entrypoints/{id}/snapshots:
    get:
      tags:
      - Entrypoints
      summary: Gets the Snapshots for the resource
      operationId: get_resource_snapshots_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the entry_point resource.
        required: true
        schema:
          type: integer
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsEntrypointPage'
  /api/v1/entrypoints/{id}/snapshots/{snapshotId}:
    get:
      tags:
      - Entrypoints
      summary: Gets a Snapshot for the resource by snapshot id
      operationId: get_resources_snapshots_id_endpoint
      parameters:
      - name: snapshotId
        in: path
        description: Snapshot ID for the entry_point resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the entry_point resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsIdEntrypoint-dump'
...
  /api/v1/experiments/{id}/snapshots:
    get:
      tags:
      - Experiments
      summary: Gets the Snapshots for the resource
      operationId: get_resource_snapshots_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the experiment resource.
        required: true
        schema:
          type: integer
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsExperimentPage'
  /api/v1/experiments/{id}/snapshots/{snapshotId}:
    get:
      tags:
      - Experiments
      summary: Gets a Snapshot for the resource by snapshot id
      operationId: get_resources_snapshots_id_endpoint
      parameters:
      - name: snapshotId
        in: path
        description: Snapshot ID for the experiment resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the experiment resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsIdExperiment-dump'
...
  /api/v1/jobs/{id}/snapshots:
    get:
      tags:
      - Job
      summary: Gets the Snapshots for the resource
      operationId: get_resource_snapshots_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the job resource.
        required: true
        schema:
          type: integer
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsJobPage'
  /api/v1/jobs/{id}/snapshots/{snapshotId}:
    get:
      tags:
      - Job
      summary: Gets a Snapshot for the resource by snapshot id
      operationId: get_resources_snapshots_id_endpoint
      parameters:
      - name: snapshotId
        in: path
        description: Snapshot ID for the job resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the job resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsIdJob-dump'
...
  /api/v1/models/{id}/snapshots:
    get:
      tags:
      - Models
      summary: Gets the Snapshots for the resource
      operationId: get_resource_snapshots_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the ml_model resource.
        required: true
        schema:
          type: integer
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsModelPage'
  /api/v1/models/{id}/snapshots/{snapshotId}:
    get:
      tags:
      - Models
      summary: Gets a Snapshot for the resource by snapshot id
      operationId: get_resources_snapshots_id_endpoint
      parameters:
      - name: snapshotId
        in: path
        description: Snapshot ID for the ml_model resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the ml_model resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsIdModel-dump'
...
  /api/v1/pluginParameterTypes/{id}/snapshots:
    get:
      tags:
      - PluginParameterTypes
      summary: Gets the Snapshots for the resource
      operationId: get_resource_snapshots_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the plugin_task_parameter_type resource.
        required: true
        schema:
          type: integer
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsPluginParameterTypePage'
  /api/v1/pluginParameterTypes/{id}/snapshots/{snapshotId}:
    get:
      tags:
      - PluginParameterTypes
      summary: Gets a Snapshot for the resource by snapshot id
      operationId: get_resources_snapshots_id_endpoint
      parameters:
      - name: snapshotId
        in: path
        description: Snapshot ID for the plugin_task_parameter_type resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin_task_parameter_type resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsIdPluginParameterType-dump'
...
  /api/v1/plugins/{id}/files/{fileId}/snapshots:
    get:
      tags:
      - Plugins
      summary: Gets the Snapshots for the resource
      operationId: get_resource_snapshots_endpoint
      parameters:
      - name: fileId
        in: path
        description: ID for the plugin_file.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin_file.
        required: true
        schema:
          type: integer
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NestedSnapshotsPluginFilePage'
  /api/v1/plugins/{id}/files/{fileId}/snapshots/{snapshotId}:
    get:
      tags:
      - Plugins
      summary: Gets a Snapshot for the resource by snapshot id
      operationId: get_resources_snapshots_id_endpoint
      parameters:
      - name: snapshotId
        in: path
        description: ID for the Snapshot.
        required: true
        schema:
          type: integer
      - name: fileId
        in: path
        description: ID for the plugin_file.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin_file.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NestedSnapshotsIdPluginFile-dump'
...
  /api/v1/plugins/{id}/snapshots:
    get:
      tags:
      - Plugins
      summary: Gets the Snapshots for the resource
      operationId: get_resource_snapshots_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the plugin resource.
        required: true
        schema:
          type: integer
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsPluginPage'
  /api/v1/plugins/{id}/snapshots/{snapshotId}:
    get:
      tags:
      - Plugins
      summary: Gets a Snapshot for the resource by snapshot id
      operationId: get_resources_snapshots_id_endpoint
      parameters:
      - name: snapshotId
        in: path
        description: Snapshot ID for the plugin resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsIdPlugin-dump'
...
  /api/v1/queues/{id}/snapshots:
    get:
      tags:
      - Queues
      summary: Gets the Snapshots for the resource
      operationId: get_resource_snapshots_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the queue resource.
        required: true
        schema:
          type: integer
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsQueuePage'
  /api/v1/queues/{id}/snapshots/{snapshotId}:
    get:
      tags:
      - Queues
      summary: Gets a Snapshot for the resource by snapshot id
      operationId: get_resources_snapshots_id_endpoint
      parameters:
      - name: snapshotId
        in: path
        description: Snapshot ID for the queue resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the queue resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsIdQueue-dump'
...
components:
  schemas:
...
    SnapshotsArtifactPage:
      type: object
      properties:
        index:
          type: integer
          description: Index of the current page.
        isComplete:
          type: boolean
          description: Boolean indicating if more data is available.
        totalNumResults:
          type: integer
          description: Total number of results.
        first:
          type: string
          description: URL to first page in results set.
        next:
          type: string
          description: URL to next page in results set.
        prev:
          type: string
          description: URL to last page in results set.
        data:
          type: array
          items:
            type: object
            description: List of Artifact resources in the current page.
            allOf:
            - $ref: '#/components/schemas/Artifact-dump'
...
    SnapshotsEntrypointPage:
      type: object
      properties:
        index:
          type: integer
          description: Index of the current page.
        isComplete:
          type: boolean
          description: Boolean indicating if more data is available.
        totalNumResults:
          type: integer
          description: Total number of results.
        first:
          type: string
          description: URL to first page in results set.
        next:
          type: string
          description: URL to next page in results set.
        prev:
          type: string
          description: URL to last page in results set.
        data:
          type: array
          items:
            type: object
            description: List of Entrypoint resources in the current page.
            allOf:
            - $ref: '#/components/schemas/Entrypoint-dump'
...